refactor(mcp): move MCP modules into src/main/mcp/ directory
