+++ src/main/mcp/mcpLaunchResolution.ts
