+++ src/main/mcp/mcpLaunchResolverManager.test.ts
